backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pillow==10.0.0
tensorflow==2.13.0
nudenet==3.0.8
opencv-python==4.8.1.78
numpy==1.24.3
python-multipart==0.0.6
requests==2.31.0
pytesseract==0.3.10
scikit-image==0.21.0
torch==2.0.1 
torchvision==0.15.2
ultralytics==8.0.186
shap==0.43.0
aiohttp==3.8.5
easyocr==1.7.0
transformers==4.34.0
openai==1.3.0
whisper==1.0.0
gtts==2.3.2
elevenlabs==0.2.21